sort js files
