+++ js/data.js
@@ -1,0 +1,86 @@
+import { getRandomInteger } from './util.js';
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = [
+  'Иван',
+  'Хуан',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Женя',
+  'Люcита',
+  'Вашингтон',
+];
+
+const DESCRIPTION = [
+  'Чудесная игра света и тени создает удивительный эффект на этой фотографии.',
+  'Кажется, будто время замерло в этой захватывающей сцене.',
+  'Цвета на этой фотографии настолько насыщены, что словно выпрыгивают из кадра.',
+  'Снимок оставляет ощущение, что мир вокруг нас необычайно прекрасен.',
+  'Невероятно, какое разнообразие форм и текстур можно увидеть на этой фотографии.',
+  'Эта фотография удивительно передает атмосферу таинственности и загадочности.',
+  'Снимок захватывает сразу несколько историй и эмоций.',
+  'Композиция на этой фотографии просто идеальна.',
+  'Фотография дает возможность увидеть мир в новом свете.',
+  'Я не могу перестать смотреть на эту фотографию - она настолько красива!',
+  'Чувствуется, как фотограф настоящий мастер своего дела.',
+  'На этой фотографии можно заметить детали, которые никогда не увидишь в обычной жизни.',
+  'Снимок прекрасно передает настроение момента и погружает в него с головой.',
+  'Фотография словно оживает и начинает рассказывать свою историю.',
+  'Такой кадр не может не вызвать улыбку на лице.',
+  'В этой фотографии присутствует неописуемое очарование.',
+  'Снимок говорит гораздо больше, чем просто тысячи слов.',
+  'Кажется, что снять такую фотографию мог бы только настоящий гений.',
+  'Снимок наполнен эмоциями, которые передаются напрямую зрителю.',
+  'В этой фотографии скрыты множество мелких деталей, которые можно заметить только при более пристальном взгляде.',
+  'Кадр словно переносит нас в другое место и время.',
+  'Снимок создает ощущение, что мир вокруг нас прекрасен в любое время года.',
+  'Кажется, что каждая фотография - это маленький мир.',
+  'Эта фотография как будто заставляет нас забыть о повседневности и наслаждаться прекрасным моментом.',
+  'Невероятно, как эта фотография передает эмоции и заставляет нас вживаться в ситуацию.',
+];
+
+let postId = 1;
+let сommentId = 1;
+
+const createComment = () => {
+  const randomComments = getRandomInteger(0, COMMENTS.length - 1);
+  const randomAvatarNumber = getRandomInteger(1, 6);
+  const randomNameIndex = getRandomInteger(0, NAMES.length - 1);
+
+  return {
+    id: сommentId++,
+    avatar: `img/avatar-${randomAvatarNumber}.svg`,
+    message: COMMENTS[randomComments],
+    name: NAMES[randomNameIndex],
+  };
+};
+
+const createPost = () => {
+  const randomUrl = getRandomInteger(1, 25);
+  const randomDescription = getRandomInteger(0, 25);
+  const randomLikes = getRandomInteger(15, 200);
+  const randomPostComments = getRandomInteger(1, 5);
+
+  return {
+    id: postId++,
+    url: `photos/${randomUrl}.jpg`,
+    description: DESCRIPTION[randomDescription],
+    likes: randomLikes,
+    comments: Array.from({length: randomPostComments}, createComment),
+  };
+};
+
+const createPosts = () => Array.from({length: 25}, createPost);
+
+export { createPosts };
+
+
